backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "product shop backend"
authors = ["hamub <hbutau35@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.1.2"
django-ninja = "^0.8.2"
flit = "^3.0.0"
psycopg2 = "^2.8.6"
black = "^20.8b1"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
